iss: Cosimulation: Moved ipc logic to a separate file
